Rename and move parseSnippetIndex to transformer file
Update references.
